move wgsl-syntax into wgsl-types
