+++ crates/wgsl-parse/Cargo.toml
@@ -10,7 +10,6 @@ license.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-wgsl-syntax = { workspace = true }
 annotate-snippets = "0.11.4"
 derive_more = { version = "2.0.1", features = [
     "as_ref",
@@ -27,6 +26,7 @@ lalrpop-util = "0.22.1"
 lexical = { version = "7.0.4", features = ["format", "power-of-two"] }
 logos = "0.15.0"
 thiserror = "2.0.12"
+wgsl-types = { workspace = true }
 
 serde = { version = "1.0.215", optional = true, features = ["derive", "rc"] }
 tokrepr = { workspace = true, optional = true }
@@ -63,11 +63,11 @@ imports = []
 # * https://github.com/gfx-rs/wgpu/issues/4410
 # * https://github.com/gfx-rs/wgpu/blob/trunk/docs/api-specs/ray_tracing.md
 # * https://github.com/gfx-rs/wgpu/blob/b93b55920a978ef9f013efe8d75cb10e69488629/naga/src/valid/mod.rs#L83
-naga_ext = ["push_constant", "wgsl-syntax/naga_ext"]
+naga_ext = ["push_constant", "wgsl-types/naga_ext"]
 push_constant = []
 
-serde = ["dep:serde", "wgsl-syntax/serde"]
-tokrepr = ["dep:tokrepr", "wgsl-syntax/tokrepr"]
+serde = ["dep:serde", "wgsl-types/serde"]
+tokrepr = ["dep:tokrepr", "wgsl-types/tokrepr"]
 
 [lints]
 workspace = true
