--- crates/wgsl-syntax/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "wgsl-syntax"
-description = "Basic representations of WGSL syntactic elements"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-version.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-serde = { version = "1.0.215", optional = true, features = ["derive"] }
-tokrepr = { workspace = true, optional = true }
-
-[features]
-default = []
-
-naga_ext = []
-serde = ["dep:serde"]
-tokrepr = ["dep:tokrepr"]
-
-[lints]
-workspace = true
